Clean up of conftest.py
