+++ tests/fixtures/__init__.py
@@ -1,0 +1,10 @@
+#!python3
+# -*- coding: utf-8 -*-
+
+"""
+Various fixtures for unit testing usage, such as example graphs.
+"""
+
+# i hate doing this so much
+from .textbook_graphs import *
+from .standard_graphs import *
